+++ CHANGELOG.md
@@ -1,9 +1,20 @@
-# Change Log
+# Changelog
 
-All notable changes to the "repototxt" extension will be documented in this file.
+All notable changes to this project will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.1.0] - 2024-10-20
 
-## [Unreleased]
+### Added
 
-- Initial release
+- Initial release of **Simple Repo to Text** extension.
+- Added functionality to convert the current workspace files into a formatted plain text document.
+- Integrated command `Simple Repo To Text - Convert Workspace to PlainText` to allow file selection from the workspace.
+- Display formatted directory structure and file contents in the generated output.
+
+### Fixed
+
+- N/A - Initial release.
+
+### Changed
+
+- N/A - Initial release.